Added mean and std dev test case

+++ ql/math/randomnumbers/xoshiro256starstarprng.cpp
@@ -22,13 +22,13 @@
 
 namespace QuantLib {
 
-    Xoshiro256StarStar::Xoshiro256StarStar(unsigned long seed)
+    Xoshiro256StarStar::Xoshiro256StarStar(unsigned long long seed)
     : Xoshiro256StarStar(seed, seed, seed, seed) {}
 
-    Xoshiro256StarStar::Xoshiro256StarStar(unsigned long s0,
-                                           unsigned long s1,
-                                           unsigned long s2,
-                                           unsigned long s3) {
+    Xoshiro256StarStar::Xoshiro256StarStar(unsigned long long s0,
+                                           unsigned long long s1,
+                                           unsigned long long s2,
+                                           unsigned long long s3) {
         seedInitialization(s0, s1, s2, s3);
 
         // using a seed needs some warm up
@@ -37,10 +37,10 @@ namespace QuantLib {
         }
     }
 
-    void Xoshiro256StarStar::seedInitialization(unsigned long s0,
-                                                unsigned long s1,
-                                                unsigned long s2,
-                                                unsigned long s3) {
+    void Xoshiro256StarStar::seedInitialization(unsigned long long s0,
+                                                unsigned long long s1,
+                                                unsigned long long s2,
+                                                unsigned long long s3) {
         if (s0 == 0 && s1 == 0 && s2 == 0 && s3 == 0) {
             s[0] = SeedGenerator::instance().get();
             s[1] = SeedGenerator::instance().get();
